# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "安然画室 · 课程预约小程序 API",
    "version": "1.0.0",
    "description": "线下课程介绍+预约微信小程序后端接口文档。包含用户端公开接口、用户端需登录接口、管理后台接口三部分。\n\n**认证方式**: JWT Bearer Token（Header: `Authorization: Bearer <token>`）\n\n**管理后台**: 用户名密码登录，使用 nickmame 作为用户名，默认密码为环境变量 `ADMIN_PASSWORD`（默认 `admin123`）\n\n**响应格式**: 统一为 `{ code: 0, message: \"ok\", data: {} }`\n\n**开发测试账号**: openid=dev_test_openid，无需真实微信code即可登录"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "本地开发环境"
    },
    {
      "url": "https://api.anran-studio.com/api",
      "description": "生产环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Token，登录后获取。用户端使用微信code登录，管理端使用用户名密码登录。"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误码"
          },
          "message": {
            "type": "string",
            "description": "错误描述"
          },
          "data": {
            "type": "null",
            "description": "始终为 null"
          }
        }
      },
...
      "CourseSchedule": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "course_type_id": {
            "type": "integer"
          },
          "date": {
            "type": "string",
            "format": "date",
            "description": "上课日期"
          },
          "start_time": {
            "type": "string",
            "description": "开始时间（HH:mm:ss）"
          },
          "end_time": {
            "type": "string",
            "description": "结束时间（HH:mm:ss）"
          },
          "max_count": {
            "type": "integer",
            "description": "最大容量"
          },
          "booked_count": {
            "type": "integer",
            "description": "已预约人数"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "cancelled",
              "full"
            ],
            "description": "时段状态"
          },
          "check_in_code": {
            "type": "string",
            "nullable": true,
            "description": "核销码（预留）"
          },
          "checked_in_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CourseType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "course_id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "description": "班型名称（如：启蒙班/进阶班）"
          },
          "capacity": {
            "type": "integer",
            "description": "班级容量"
          },
          "price": {
            "type": "number",
            "description": "价格（预留）"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Student": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "description": "学员姓名"
          },
          "age": {
            "type": "integer",
            "description": "年龄"
          },
          "phone": {
            "type": "string",
            "description": "联系电话"
          },
          "notes": {
            "type": "string",
            "description": "备注"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "OrderSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "course_id": {
            "type": "integer"
          },
          "course_type_id": {
            "type": "integer"
          },
          "schedule_id": {
            "type": "integer"
          },
          "student_id": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "cancelled"
            ],
            "description": "订单状态"
          },
          "audit_remark": {
            "type": "string",
            "nullable": true,
            "description": "审核备注"
          },
          "pay_status": {
            "type": "string",
            "enum": [
              "unpaid",
              "paid",
              "refunded"
            ],
            "description": "支付状态（预留）"
          },
          "transaction_id": {
            "type": "string",
            "nullable": true
          },
          "course": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "title": {
                "type": "string"
              },
              "cover": {
                "type": "string"
              }
            }
          },
          "courseType": {
            "$ref": "#/components/schemas/CourseType"
          },
          "schedule": {
            "$ref": "#/components/schemas/CourseSchedule"
          },
          "student": {
            "$ref": "#/components/schemas/Student"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "OrderDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/OrderSummary"
          },
          {
            "type": "object",
            "properties": {
              "course": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "title": {
                    "type": "string"
                  },
                  "cover": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ]
      },
      "OrderAdmin": {
        "allOf": [
          {
            "$ref": "#/components/schemas/OrderSummary"
          },
          {
            "type": "object",
            "properties": {
              "user": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "nickname": {
                    "type": "string"
                  },
                  "avatar": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ]
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "未登录或 Token 无效",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "noToken": {
                "value": {
                  "code": 401,
                  "message": "请先登录",
                  "data": null
                }
              },
              "expired": {
                "value": {
                  "code": 401,
                  "message": "登录已过期，请重新登录",
                  "data": null
                }
              },
              "invalid": {
                "value": {
                  "code": 401,
                  "message": "无效的登录凭证",
                  "data": null
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "无管理员权限",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "code": 403,
              "message": "无管理员权限",
              "data": null
            }
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/orders": {
      "post": {
        "tags": [
          "预约订单"
        ],
        "summary": "提交预约",
        "description": "用户提交课程预约。检查时段名额（行级锁防超卖）和同一用户重复提交。成功后booked_count+1。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "course_id",
                  "course_type_id",
                  "schedule_id",
                  "student_id"
                ],
                "properties": {
                  "course_id": {
                    "type": "integer",
                    "description": "课程ID"
                  },
                  "course_type_id": {
                    "type": "integer",
                    "description": "班型ID"
                  },
                  "schedule_id": {
                    "type": "integer",
                    "description": "时段ID"
                  },
                  "student_id": {
                    "type": "integer",
                    "description": "学员ID"
                  }
                }
              },
              "example": {
                "course_id": 1,
                "course_type_id": 1,
                "schedule_id": 1,
                "student_id": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "预约提交成功，状态为pending",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "预约已提交，请等待管理员审核"
                    },
                    "data": {
                      "anyOf": [
                        {
                          "$ref": "#/components/schemas/OrderSummary"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数缺失/名额已满/重复提交",
            "content": {
              "application/json": {
                "examples": {
                  "missingParams": {
                    "value": {
                      "code": 400,
                      "message": "缺少必填参数",
                      "data": null
                    }
                  },
                  "duplicate": {
                    "value": {
                      "code": 400,
                      "message": "该时段您已有预约，请勿重复提交",
                      "data": null
                    }
                  },
                  "cancelled": {
                    "value": {
                      "code": 400,
                      "message": "该时段已取消",
                      "data": null
                    }
                  },
                  "full": {
                    "value": {
                      "code": 400,
                      "message": "名额已满",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "get": {
        "tags": [
          "预约订单"
        ],
        "summary": "我的预约列表",
        "description": "获取当前用户的预约订单列表，支持按状态筛选和分页。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页条数",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "按状态筛选（不传返回全部）",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected",
                "cancelled"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页订单列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "list": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/OrderSummary"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "pageSize": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      },
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "tags": [
          "预约订单"
        ],
        "summary": "预约详情",
        "description": "获取当前用户指定订单的详细信息。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "订单ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "订单完整信息",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "anyOf": [
                        {
                          "$ref": "#/components/schemas/OrderDetail"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "订单不存在",
            "content": {
              "application/json": {
                "example": {
                  "code": 404,
                  "message": "订单不存在",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}/cancel": {
      "put": {
        "tags": [
          "预约订单"
        ],
        "summary": "取消预约",
        "description": "用户取消自己的预约（仅pending或approved状态可取消），取消后释放名额（booked_count-1）。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "订单ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "取消成功",
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "message": "已取消预约",
                  "data": null
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "nullable": true,
                      "description": "操作类接口成功时返回 null"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "当前状态不可取消",
            "content": {
              "application/json": {
                "example": {
                  "code": 400,
                  "message": "当前状态不可取消",
                  "data": null
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "订单不存在",
            "content": {
              "application/json": {
                "example": {
                  "code": 404,
                  "message": "订单不存在",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/orders": {
      "get": {
        "tags": [
          "管理端-订单"
        ],
        "summary": "预约订单列表",
        "description": "获取所有用户的预约订单，含用户、课程、班型、时段、学员信息。支持按状态筛选和分页。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "按状态筛选",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected",
                "cancelled"
              ]
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "description": "搜索（目前仅用于前端过滤，后端未实现搜索逻辑）",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页订单列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "list": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/OrderAdmin"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "pageSize": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      },
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/admin/orders/{id}": {
      "get": {
        "tags": [
          "管理端-订单"
        ],
        "summary": "订单详情",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "订单ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "订单完整信息",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "ok"
                    },
                    "data": {
                      "anyOf": [
                        {
                          "$ref": "#/components/schemas/OrderAdmin"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "订单不存在",
            "content": {
              "application/json": {
                "example": {
                  "code": 404,
                  "message": "订单不存在",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/orders/{id}/cancel": {
      "put": {
        "tags": [
          "管理端-订单"
        ],
        "summary": "管理员取消订单",
        "description": "管理员强制取消订单（任意状态）。取消后释放名额。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "订单ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "已取消",
            "content": {
              "application/json": {
                "example": {
                  "code": 0,
                  "message": "已取消",
                  "data": null
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "nullable": true,
                      "description": "操作类接口成功时返回 null"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "订单已取消",
            "content": {
              "application/json": {
                "example": {
                  "code": 400,
                  "message": "订单已取消",
                  "data": null
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "订单不存在",
            "content": {
              "application/json": {
                "example": {
                  "code": 404,
                  "message": "订单不存在",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
...
  }
}